Add System/Callback prioritization
