+++ engine/Assets/Scripts/Engine/ModuleLoader/Api.cs
@@ -26,7 +26,6 @@ using Engine.ModuleLoader.Adapters;
 using Logger = SynthesisAPI.Utilities.Logger;
 using Directory = System.IO.Directory;
 
-using PreloadedModule = System.ValueTuple<System.IO.Compression.ZipArchive, Engine.ModuleLoader.ModuleMetadata>;
 
 namespace Engine.ModuleLoader
 {
@@ -51,13 +50,13 @@ namespace Engine.ModuleLoader
 
 			try
 			{
-				LoadApi();
+				ModuleLoader.PreloadApi();
 			}
 			catch (Exception e)
 			{
 				Logger.Log($"Failed to load API\n{e}", LogLevel.Error);
 			}
-			LoadModules();
+			ModuleLoader.Run(ModulesSourcePath, BaseModuleTargetPath);
 			RerouteConsoleOutput();
 		}
 
@@ -66,388 +65,7 @@ namespace Engine.ModuleLoader
 			ToastLogger.ScrollToBottom();
 		}
 
-		private void LoadModules()
-		{
-			if (!Directory.Exists(ModulesSourcePath))
-			{
-				Directory.CreateDirectory(ModulesSourcePath);
-			}
-			List<PreloadedModule> modules = new List<PreloadedModule>();
-			try
-			{
-				modules = PreloadModules();
-			}
-			catch (Exception e)
-			{
-				Logger.Log($"Failed to preload modules\n{e}", LogLevel.Error);
-			}
-			try
-			{
-				ResolveDependencies(modules);
-			}
-			catch (Exception e)
-			{
-				Logger.Log($"Failed to resolve module dependencies\n{e}", LogLevel.Error);
-			}
-			foreach (var (archive, metadata) in modules)
-			{
-				try
-				{
-					LoadModule((archive, metadata));
-
-					EventBus.Push(new LoadModuleEvent(metadata.Name, metadata.Version));
-					ModuleManager.AddToLoadedModuleList(new ModuleManager.ModuleInfo(metadata.Name, metadata.Version));
-				}
-				catch (Exception e)
-				{
-					Logger.Log($"Failed to load module {metadata.Name}\n{e}", LogLevel.Error);
-					// TODO error screen
-					throw e;
-				}
-			}
-			foreach (var (archive, _) in modules)
-			{
-				archive.Dispose();
-			}
-			ModuleManager.MarkFinishedLoading();
-		}
-
-		private List<PreloadedModule> PreloadModules()
-		{
-			var modules = new List<PreloadedModule>();
-
-			// Discover and preload all modules
-			foreach (var file in Directory.GetFiles(ModulesSourcePath).Where(fn => Path.GetExtension(fn) == ".zip"))
-			{
-				var module = PreloadModule(Path.GetFileName(file));
-				if (module is null)
-					continue;
-				foreach (var (_, metadata) in modules)
-				{
-					if (metadata.Name == module?.Item2.Name)
-					{
-						foreach (var (archive, _) in modules)
-						{
-							archive.Dispose();
-						}
-						throw new LoadModuleException($"Attempting to load module with duplicate name: {metadata.Name}");
-					}
-					if (metadata.TargetPath == module?.Item2.TargetPath)
-					{
-						foreach (var (archive, _) in modules)
-						{
-							archive.Dispose();
-						}
-						throw new LoadModuleException($"Attempting to load modules into same target path: {metadata.TargetPath}");
-					}
-				}
-				modules.Add(module.Value);
-			}
-			return modules;
-		}
-
-		private PreloadedModule? PreloadModule(string filePath)
-		{
-			var fullPath = $"{ModulesSourcePath}{Path.DirectorySeparatorChar}{filePath}";
-
-			var module = ZipFile.Open(fullPath, ZipArchiveMode.Read);
-
-			// Ensure module contains metadata
-			if (module.Entries.All(e => e.Name != ModuleMetadata.MetadataFilename))
-			{
-				Logger.Log($"Potential module missing is metadata file: {filePath}", LogLevel.Warning);
-				return null;
-			}
-
-			// Parse module metadata
-			try
-			{
-				var metadata = ModuleMetadata.Deserialize(module.Entries
-					.First(e => e.Name == ModuleMetadata.MetadataFilename).Open());
-				return (module, metadata);
-			}
-			catch (Exception e)
-			{
-				throw new LoadModuleException($"Failed to deserialize metadata in module: {fullPath}", e);
-			}
-		}
-
-		private void ResolveDependencies(List<(ZipArchive archive, ModuleMetadata metadata)> moduleList)
-		{
-			foreach (var (_, metadata) in moduleList)
-			{
-				foreach (var dependency in metadata.Dependencies)
-				{
-					if (moduleList.All(m => m.metadata.Name != dependency.Name))
-					{
-						throw new LoadModuleException($"Module {metadata.Name} is missing dependency module {dependency.Name}");
-					}
-					var present_dep = moduleList.First(m => m.metadata.Name == dependency.Name);
-					if (present_dep.metadata.Version != dependency.Version)
-					{
-						throw new LoadModuleException($"Module {metadata.Name} requires dependency module {dependency.Name} version {dependency.Version} but its version is {present_dep.metadata.Version}");
-					}
-				}
-			}
-
-			// Use Kahns algorithm to resolve module dependencies, ordering modules in list
-			// in the order they should be loaded
-
-			// TODO check for cyclic dependencies and throw
-			var resolvedEntries = moduleList.Where(t => !t.metadata.Dependencies.Any()).ToList();
-			var solutionSet = new Queue<(ZipArchive archive, ModuleMetadata metadata)>();
-			while (resolvedEntries.Count > 0)
-			{
-				var element = resolvedEntries.PopAt(0);
-				solutionSet.Enqueue(element);
-				foreach (var dep in moduleList.Where(t =>
-					t.metadata.Dependencies.Any(d => d.Name == element.metadata.Name && d.Version == element.metadata.Version)).ToList())
-				{
-					dep.metadata.Dependencies.RemoveAll(d => d.Name == element.metadata.Name && d.Version == element.metadata.Version);
-					if (dep.metadata.Dependencies.Count == 0)
-						resolvedEntries.Add(dep);
-				}
-			}
-
-			moduleList.Clear();
-			moduleList.AddRange(solutionSet.ToList());
-		}
-
-		private static string GetPath(string fullName)
-		{
-			var i = fullName.LastIndexOf(SynthesisAPI.VirtualFileSystem.Directory.DirectorySeparatorChar, fullName.Length - 1, fullName.Length - 2);
-			if (i == -1 || i == (fullName.Length - 1))
-			{
-				return "";
-			}
-			return fullName.Substring(0, i + 1);
-		}
-
-		private static string RemovePath(string metadataPath, string fullName)
-		{
-			if (fullName.StartsWith(metadataPath))
-			{
-				return fullName.Substring(metadataPath.Length);
-			}
-			return fullName;
-		}
-
-		private void LoadModule((ZipArchive archive, ModuleMetadata metadata) moduleInfo)
-		{
-			var fileManifest = new List<string>();
-			fileManifest.AddRange(moduleInfo.metadata.FileManifest);
-
-			var metadataPath = GetPath(moduleInfo.archive.Entries.First(e => e.Name == ModuleMetadata.MetadataFilename).FullName);
-
-			foreach (var entry in moduleInfo.archive.Entries.Where(e =>
-			{
-				var name = RemovePath(metadataPath, e.FullName);
-				return name != ModuleMetadata.MetadataFilename && moduleInfo.metadata.FileManifest.Contains(name);
-			}))
-			{
-				fileManifest.Remove(RemovePath(metadataPath, entry.FullName));
-				var extension = Path.GetExtension(entry.Name);
-				var stream = entry.Open();
-				if (extension == ".dll")
-				{
-					if (!LoadModuleAssembly(stream, moduleInfo.metadata.Name))
-					{
-						throw new LoadModuleException($"Failed to load assembly: {entry.Name}");
-					}
-				}
-				else
-				{
-					var targetPath = BaseModuleTargetPath + SynthesisAPI.VirtualFileSystem.Directory.DirectorySeparatorChar +
-						moduleInfo.metadata.TargetPath + SynthesisAPI.VirtualFileSystem.Directory.DirectorySeparatorChar +
-						GetPath(RemovePath(metadataPath, entry.FullName));
-					var perm = Permissions.PublicReadWrite;
-					var type = AssetManager.GetTypeFromFileExtension(extension);
-					if (type == null)
-					{
-						throw new LoadModuleException($"Failed to determine asset type from file extension of asset: {entry.Name}");
-					}
-					else if (AssetManager.Import(type,
-						new DeflateStreamWrapper(stream, entry.Length), targetPath, entry.Name, perm, "") == null)
-					{
-						throw new LoadModuleException($"Failed to import asset: {entry.Name}");
-					}
-				}
-				// Debug.Log("Loaded " + entry.Name);
-			}
-			foreach (var file in fileManifest)
-			{
-				Logger.Log($"Module \"{moduleInfo.metadata.Name}\" is missing file from manifest: {file}", LogLevel.Warning);
-			}
-		}
-
-		private bool LoadApi()
-		{
-			// Set up Api
-			try
-			{
-				var apiAssembly = AppDomain.CurrentDomain.GetAssemblies().First(a => a.GetName().Name == "Api");
-				foreach (var type in apiAssembly.GetTypes()
-					.Where(t => t.GetCustomAttribute<ModuleOmitAttribute>() == null &&
-					(t.IsSubclassOf(typeof(SystemBase)) ||
-						t.GetMethods().Any(m => m.GetCustomAttribute<CallbackAttribute>() != null || 
-						m.GetCustomAttribute<TaggedCallbackAttribute>() != null))))
-				{
-					object instance = null;
-
-					if (type.IsSubclassOf(typeof(SystemBase)))
-					{
-						var entity = EnvironmentManager.AddEntity();
-						instance = entity.AddComponent(type);
-					}
-
-					if (instance == null && type.GetMethods().Any(m =>
-						m.GetCustomAttribute<CallbackAttribute>() != null ||
-						m.GetCustomAttribute<TaggedCallbackAttribute>() != null))
-					{
-						instance = Activator.CreateInstance(type);
-					}
-
-					foreach (var callback in type.GetMethods()
-						.Where(m => m.GetCustomAttribute<CallbackAttribute>() != null))
-					{
-						RegisterTypeCallback(callback, instance);
-					}
-					foreach (var callback in type.GetMethods()
-						.Where(m => m.GetCustomAttribute<TaggedCallbackAttribute>() != null))
-					{
-						RegisterTagCallback(callback, instance);
-					}
-				}
-				ModuleManager.RegisterModuleAssemblyName(apiAssembly.GetName().Name, "Api");
-			}
-			catch (Exception e)
-			{
-				Logger.Log($"Failed to load API\n{e}", LogLevel.Error);
-			}
-			return true;
-		}
-
-		private bool LoadModuleAssembly(Stream stream, string owningModule)
-		{
-			// Load module assembly
-			var memStream = new MemoryStream();
-			stream.CopyTo(memStream);
-			stream.Close();
-			var assembly = Assembly.Load(memStream.ToArray());
-
-			// Set up module
-			Type[] types;
-			try
-			{
-				types = assembly.GetTypes();
-			}
-			catch (ReflectionTypeLoadException e)
-			{
-				if (e is ReflectionTypeLoadException reflectionTypeLoadException)
-				{
-					foreach (var inner in reflectionTypeLoadException.LoaderExceptions)
-					{
-						Logger.Log($"Loading module {owningModule} resulted in type errors\n{inner}", LogLevel.Error);
-					}
-				}
-				return false;
-			}
-			catch (Exception e)
-			{
-				Logger.Log($"Failed to get types from module {owningModule}\n{e}", LogLevel.Error);
-				return false;
-			}
-
-			foreach (var exportedModuleClass in types
-				.Where(t => t.GetCustomAttribute<ModuleOmitAttribute>() == null && 
-					(t.IsSubclassOf(typeof(SystemBase)) ||
-						t.GetMethods().Any(m => m.GetCustomAttribute<CallbackAttribute>() != null
-						|| m.GetCustomAttribute<TaggedCallbackAttribute>() != null))))
-			{
-				try
-				{
-					object exportedModuleClassInstance = null;
-
-					if (exportedModuleClass.IsSubclassOf(typeof(SystemBase)))
-					{
-						var entity = EnvironmentManager.AddEntity();
-						exportedModuleClassInstance = entity.AddComponent(exportedModuleClass);
-					}
-
-					if(exportedModuleClassInstance == null)
-						exportedModuleClassInstance = Activator.CreateInstance(exportedModuleClass);
-
-					foreach (var callback in exportedModuleClass.GetMethods()
-						.Where(m => m.GetCustomAttribute<CallbackAttribute>() != null))
-					{
-						RegisterTypeCallback(callback, exportedModuleClassInstance);
-					}
-
-					foreach (var callback in exportedModuleClass.GetMethods()
-						.Where(m => m.GetCustomAttribute<TaggedCallbackAttribute>() != null))
-					{
-						RegisterTagCallback(callback, exportedModuleClassInstance);
-					}
-				}
-				catch (TypeLoadException e)
-				{
-					Logger.Log($"Module loader failed to process type {exportedModuleClass} from module {owningModule}\nType: {e.TypeName}, Site: {e.TargetSite}\n{e}", LogLevel.Error);
-				}
-				catch (Exception e)
-				{
-					Logger.Log($"Module loader failed to process type {exportedModuleClass} from module {owningModule}\n{e}", LogLevel.Error);
-					// TODO unload assembly? return false?
-					continue;
-				}
-			}
-
-			ModuleManager.RegisterModuleAssemblyName(assembly.GetName().Name, owningModule);
-			return true;
-		}
-
-		private void RegisterTagCallback(MethodInfo callback, object instance)
-		{
-			if (instance.GetType() != callback.DeclaringType) // Sanity check
-			{
-				throw new LoadModuleException(
-					$"Type of instance variable \"{instance.GetType()}\" does not match declaring type of callback \"{callback.Name}\" (expected \"{callback.DeclaringType}\"");
-			}
-			var eventType = callback.GetParameters().First().ParameterType;
-			var tag = callback.GetCustomAttribute<TaggedCallbackAttribute>().Tag;
-			typeof(EventBus).GetMethod("NewTagListener").Invoke(null, new object[]
-				{
-					tag,
-					CreateEventCallback(callback, instance, eventType)
-				});
-		}
-
-		private void RegisterTypeCallback(MethodInfo callback, object instance)
-		{
-			if (instance.GetType() != callback.DeclaringType) // Sanity check
-			{
-				throw new LoadModuleException(
-					$"Type of instance variable \"{instance.GetType()}\" does not match declaring type of callback \"{callback.Name}\" (expected \"{callback.DeclaringType}\"");
-			}
-			var eventType = callback.GetParameters().First().ParameterType;
-			typeof(EventBus).GetMethod("NewTypeListener")
-				?.MakeGenericMethod(eventType).Invoke(null, new object[]
-				{
-					CreateEventCallback(callback, instance, eventType)
-				});
-		}
-
-		EventBus.EventCallback CreateEventCallback(MethodInfo m, object instance, Type eventType)
-		{
-			return (e) => m.Invoke(instance,
-				new[]
-				{
-					typeof(ReflectHelper).GetMethod("CastObject")
-						?.MakeGenericMethod(eventType)
-						.Invoke(null, new object[] {e})
-				});
-		}
-
-		private static void SetupApplication()
+		private void SetupApplication()
 		{
 			Application.logMessageReceivedThreaded +=
 				(condition, stackTrace, type) =>
@@ -477,7 +95,6 @@ namespace Engine.ModuleLoader
 
 			var _ = SynthesisAPI.UIManager.UIManager.RootElement;
 		}
-
 		private class LoggerImpl : SynthesisAPI.Utilities.ILogger
 		{
 			private bool debugLogsEnabled = true;
@@ -498,10 +115,8 @@ namespace Engine.ModuleLoader
 						{
 							return;
 						}
-						{
-							Debug.Log(msg);
-							break;
-						}
+						Debug.Log(msg);
+						break;
 					case LogLevel.Warning:
 						{
 							Debug.LogWarning(msg);
