+++ engine/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.0f1
-m_EditorVersionWithRevision: 2019.4.0f1 (0af376155913)
+m_EditorVersion: 2019.4.1f1
+m_EditorVersionWithRevision: 2019.4.1f1 (e6c045e14e4e)
